+++ src/core/SkShaderCodeDictionary.cpp
@@ -29,40 +29,13 @@ using DataPayloadType = SkPaintParamsKey::DataPayloadType;
 
 namespace {
 
+#if defined(SK_GRAPHITE_ENABLED) && defined(SK_METAL)
 std::string get_mangled_name(const std::string& baseName, int manglingSuffix) {
     return baseName + "_" + std::to_string(manglingSuffix);
 }
 
 void add_indent(std::string* result, int indent) {
     result->append(4*indent, ' ');
-}
-
-#if defined(SK_GRAPHITE_ENABLED) && defined(SK_METAL)
-std::string generate_default_before_children_glue_code(int entryIndex,
-                                                       const SkPaintParamsKey::BlockReader& reader,
-                                                       const std::string& parentPreLocalName,
-                                                       int indent) {
-    std::string result;
-
-    if (reader.entry()->needsLocalCoords()) {
-        // Every snippet that requests local coordinates must have a preLocalMatrix as its first
-        // uniform
-        SkASSERT(reader.entry()->fUniforms.size() >= 1);
-        SkASSERT(reader.entry()->fUniforms[0].type() == SkSLType::kFloat4x4);
-
-        std::string localMatrixUniformName = reader.entry()->getMangledUniformName(0, entryIndex);
-
-        std::string preLocalMatrixVarName = get_mangled_name("preLocal", entryIndex);
-
-        add_indent(&result, indent);
-        SkSL::String::appendf(&result,
-                              "float4x4 %s = %s * %s;\n",
-                              preLocalMatrixVarName.c_str(),
-                              parentPreLocalName.c_str(),
-                              localMatrixUniformName.c_str());
-    }
-
-    return result;
 }
 #endif
 
@@ -89,26 +62,52 @@ std::string GetMtlTexturesAndSamplers(const std::vector<SkPaintParamsKey::BlockR
                                       int* binding);
 } // namespace skgpu::graphite
 
-// Emit the glue code needed to invoke a single static helper isolated w/in its own scope.
-// The structure of this will be:
+// Returns an expression to calculate the pre-local matrix for a given entry.
+
+static std::string pre_local_matrix_for_entry(const SkShaderInfo& shaderInfo,
+                                              int entryIndex,
+                                              const std::string& parentMatrix) {
+    const SkPaintParamsKey::BlockReader& reader = shaderInfo.blockReader(entryIndex);
+    if (!reader.entry()->needsLocalCoords()) {
+        // Return the parent matrix as-is.
+        return parentMatrix;
+    }
+
+    // The snippet requested local coordinates, so the pre-local matrix must be its first uniform.
+    SkASSERT(reader.entry()->fUniforms.size() >= 1);
+    SkASSERT(reader.entry()->fUniforms.front().type() == SkSLType::kFloat4x4);
+
+    std::string localMatrixUniformName = reader.entry()->getMangledUniformName(0, entryIndex);
+    return SkSL::String::printf("(%s * %s)", parentMatrix.c_str(), localMatrixUniformName.c_str());
+}
+
+// Emit the glue code needed to invoke a single static helper isolated within its own scope.
+// Glue code will assign the resulting color into a variable `half4 outColor%d`, where the %d is
+// filled in with 'entryIndex'.
+// Glue code is allowed to emit children recursively, which leads to a nested structure like this:
 //
-//     half4 outColor%d;
+//     half4 outColor1;  // output of shader
 //     {
-//         half4 child-outColor%d;  // for each child
+//         half4 outColor2;  // output of first child
 //         {
-//             /* emitted snippet sksl assigns to child-outColor%d */
+//             outColor2 = sk_first_child_snippet(uniformA, uniformB);
+//         }
+//         half4 outColor3;  // output of second child
+//         {
+//             outColor3 = sk_second_child_snippet(uniformC, uniformD);
 //         }
 //
-//         /* emitted snippet sksl assigns to outColor%d - taking a vector of child var names */
+//         outColor1 = sk_shader_snippet(uniformE, outColor2, outColor3);
 //     }
-// Where the %d is filled in with 'entryIndex'.
-std::string SkShaderInfo::emitGlueCodeForEntry(int* entryIndex,
-                                               const std::string& priorStageOutputName,
-                                               const std::string& parentPreLocalName,
-                                               std::string* preamble,
-                                               std::string* mainBody,
-                                               int indent) const {
-    const SkPaintParamsKey::BlockReader& reader = fBlockReaders[*entryIndex];
+
+static std::string emit_glue_code_for_entry(const SkShaderInfo& shaderInfo,
+                                            int* entryIndex,
+                                            const std::string& priorStageOutputName,
+                                            const std::string& parentPreLocalName,
+                                            std::string* preamble,
+                                            std::string* mainBody,
+                                            int indent) {
+    const SkPaintParamsKey::BlockReader& reader = shaderInfo.blockReader(*entryIndex);
     int curEntryIndex = *entryIndex;
 
     std::string scopeOutputVar = get_mangled_name("outColor", curEntryIndex);
@@ -121,36 +120,51 @@ std::string SkShaderInfo::emitGlueCodeForEntry(int* entryIndex,
     add_indent(mainBody, indent);
     *mainBody += "{\n";
 
-    *mainBody += generate_default_before_children_glue_code(curEntryIndex, reader,
-                                                            parentPreLocalName, indent + 1);
-
-    // TODO: this could be returned by generate_default_before_children_glue_code
     std::string currentPreLocalName;
     if (reader.entry()->needsLocalCoords()) {
         currentPreLocalName = get_mangled_name("preLocal", curEntryIndex);
+        std::string preLocalExpression = pre_local_matrix_for_entry(shaderInfo,
+                                                                    curEntryIndex,
+                                                                    parentPreLocalName);
+        add_indent(mainBody, indent + 1);
+        SkSL::String::appendf(mainBody,
+                              "float4x4 %s = %s;\n",
+                              currentPreLocalName.c_str(),
+                              preLocalExpression.c_str());
     } else {
+        // Inherit the parent matrix; reuse the same variable instead of introducing a new one.
         currentPreLocalName = parentPreLocalName;
     }
 
-    // Although the children appear after the parent in the shader info they are emitted
-    // before the parent
-    std::vector<std::string> childOutputVarNames;
-    for (int j = 0; j < reader.numChildren(); ++j) {
-        *entryIndex += 1;
-        std::string childOutputVar = this->emitGlueCodeForEntry(entryIndex,
-                                                                priorStageOutputName,
-                                                                currentPreLocalName,
-                                                                preamble, mainBody, indent + 1);
-        childOutputVarNames.push_back(childOutputVar);
-    }
-
-    (reader.entry()->fGlueCodeGenerator)(scopeOutputVar, curEntryIndex, reader,
-                                         fRuntimeEffectDictionary, priorStageOutputName,
-                                         childOutputVarNames, preamble, mainBody, indent + 1);
+    (reader.entry()->fGlueCodeGenerator)(shaderInfo, scopeOutputVar, entryIndex, reader,
+                                         priorStageOutputName, currentPreLocalName,
+                                         preamble, mainBody, indent + 1);
     add_indent(mainBody, indent);
     *mainBody += "}\n";
 
     return scopeOutputVar;
+}
+
+static std::vector<std::string> emit_child_glue_code(const SkShaderInfo& shaderInfo,
+                                                     int* entryIndex,
+                                                     const std::string& priorStageOutputName,
+                                                     const std::string& currentPreLocalName,
+                                                     std::string* preamble,
+                                                     std::string* mainBody,
+                                                     int indent) {
+    const SkPaintParamsKey::BlockReader& reader = shaderInfo.blockReader(*entryIndex);
+    const int numChildren = reader.numChildren();
+
+    std::vector<std::string> childOutputVarNames;
+    for (int j = 0; j < numChildren; ++j) {
+        *entryIndex += 1;
+        std::string childOutputVar = emit_glue_code_for_entry(shaderInfo, entryIndex,
+                                                              priorStageOutputName,
+                                                              currentPreLocalName,
+                                                              preamble, mainBody, indent);
+        childOutputVarNames.push_back(std::move(childOutputVar));
+    }
+    return childOutputVarNames;
 }
 
 // The current, incomplete, model for shader construction is:
@@ -187,8 +201,8 @@ std::string SkShaderInfo::toSkSL() const {
     SkSL::String::appendf(&mainBody, "    half4 %s = half4(0);", lastOutputVar.c_str());
 
     for (int entryIndex = 0; entryIndex < (int) fBlockReaders.size(); ++entryIndex) {
-        lastOutputVar = this->emitGlueCodeForEntry(&entryIndex, lastOutputVar, parentPreLocal,
-                                                   &preamble, &mainBody, 1);
+        lastOutputVar = emit_glue_code_for_entry(*this, &entryIndex, lastOutputVar, parentPreLocal,
+                                                 &preamble, &mainBody, /*indent=*/1);
     }
 
     SkSL::String::appendf(&mainBody, "    sk_FragColor = %s;\n", lastOutputVar.c_str());
@@ -301,54 +315,108 @@ void SkShaderCodeDictionary::getShaderInfo(SkUniquePaintParamsID uniqueID, SkSha
 //--------------------------------------------------------------------------------------------------
 namespace {
 
+#if defined(SK_GRAPHITE_ENABLED) && defined(SK_ENABLE_SKSL)
+static std::string append_default_snippet_arguments(const SkShaderSnippet* entry,
+                                                    int entryIndex,
+                                                    const std::string& currentPreLocalName,
+                                                    SkSpan<const std::string> childOutputs) {
+    std::string code = "(";
+
+    // Append uniform names.
+    const char* separator = "";
+    for (size_t i = 0; i < entry->fUniforms.size(); ++i) {
+        code += separator;
+        separator = ", ";
+
+        if (i == 0 && entry->needsLocalCoords()) {
+            code += currentPreLocalName + " * dev2LocalUni";
+        } else {
+            code += entry->getMangledUniformName(i, entryIndex);
+        }
+    }
+
+    // Append child output names.
+    for (const std::string& childOutputVar : childOutputs) {
+        code += separator;
+        separator = ", ";
+        code += childOutputVar;
+    }
+    code.push_back(')');
+
+    return code;
+}
+#endif
+
 // The default glue code just calls a helper function with the signature:
-//    half4 fStaticFunctionName(/* all uniforms as parameters */,
-//                              /* all child output variable names as parameters */);
+//    half4 BuiltinFunctionName(/* all uniforms as parameters */);
 // and stores the result in a variable named "resultName".
-void GenerateDefaultGlueCode(const std::string& resultName,
-                             int entryIndex,
+void GenerateDefaultGlueCode(const SkShaderInfo& shaderInfo,
+                             const std::string& resultName,
+                             int* entryIndex,
                              const SkPaintParamsKey::BlockReader& reader,
-                             const SkRuntimeEffectDictionary*,
                              const std::string& priorStageOutputName,
-                             const std::vector<std::string>& childOutputVarNames,
+                             const std::string& currentPreLocalName,
                              std::string* preamble,
                              std::string* mainBody,
                              int indent) {
+#if defined(SK_GRAPHITE_ENABLED) && defined(SK_ENABLE_SKSL)
     const SkShaderSnippet* entry = reader.entry();
-
-    SkASSERT((int)childOutputVarNames.size() == entry->fNumChildren);
+    SkASSERT(entry->fNumChildren == 0);
 
     if (entry->needsLocalCoords()) {
-        // Every snippet that requests local coordinates must have a localMatrix as its first
-        // uniform
-        SkASSERT(reader.entry()->fUniforms.size() >= 1);
-        SkASSERT(reader.entry()->fUniforms[0].type() == SkSLType::kFloat4x4);
+        // Any snippet that requests local coordinates must have a localMatrix as its first uniform.
+        SkASSERT(entry->fUniforms.size() >= 1);
+        SkASSERT(entry->fUniforms.front().type() == SkSLType::kFloat4x4);
     }
 
     add_indent(mainBody, indent);
-    SkSL::String::appendf(mainBody,
-                          "%s = %s(",
-                          resultName.c_str(),
-                          entry->fStaticFunctionName);
-    const char* separator = "";
-    for (size_t i = 0; i < entry->fUniforms.size(); ++i) {
-        *mainBody += separator;
-        separator = ", ";
+    SkSL::String::appendf(mainBody, "%s = %s", resultName.c_str(), entry->fStaticFunctionName);
+    *mainBody += append_default_snippet_arguments(entry, *entryIndex, currentPreLocalName,
+                                                  /*childOutputs=*/{});
+    *mainBody += ";\n";
+#endif  // defined(SK_GRAPHITE_ENABLED) && defined(SK_ENABLE_SKSL)
+}
 
-        if (i == 0 && reader.entry()->needsLocalCoords()) {
-            *mainBody += get_mangled_name("preLocal", entryIndex);
-            *mainBody += " * dev2LocalUni";
-        } else {
-            *mainBody += entry->getMangledUniformName(i, entryIndex);
-        }
-    }
-    for (size_t i = 0; i < childOutputVarNames.size(); ++i) {
-        *mainBody += separator;
-        separator = ", ";
+// The default-with-children glue code just calls a helper function with the signature:
+//    half4 BuiltinFunctionName(/* all uniforms as parameters */,
+//                              /* all child output variable names as parameters */);
+// and stores the result in a variable named "resultName".
+void GenerateDefaultGlueCodeWithChildren(const SkShaderInfo& shaderInfo,
+                                         const std::string& resultName,
+                                         int* entryIndex,
+                                         const SkPaintParamsKey::BlockReader& reader,
+                                         const std::string& priorStageOutputName,
+                                         const std::string& currentPreLocalName,
+                                         std::string* preamble,
+                                         std::string* mainBody,
+                                         int indent) {
+#if defined(SK_GRAPHITE_ENABLED) && defined(SK_ENABLE_SKSL)
+    const SkShaderSnippet* entry = reader.entry();
+    SkASSERT(entry->fNumChildren > 0);
 
-        *mainBody += childOutputVarNames[i];
+    if (entry->needsLocalCoords()) {
+        // Any snippet that requests local coordinates must have a localMatrix as its first uniform.
+        SkASSERT(entry->fUniforms.size() >= 1);
+        SkASSERT(entry->fUniforms.front().type() == SkSLType::kFloat4x4);
     }
-    *mainBody += ");\n";
+
+    // Invoke all children ahead of this entry's glue code.
+    int curEntryIndex = *entryIndex;
+    std::vector<std::string> childOutputVarNames = emit_child_glue_code(shaderInfo,
+                                                                        entryIndex,
+                                                                        priorStageOutputName,
+                                                                        currentPreLocalName,
+                                                                        preamble,
+                                                                        mainBody,
+                                                                        indent);
+    SkASSERT((int)childOutputVarNames.size() == entry->fNumChildren);
+
+    add_indent(mainBody, indent);
+    SkSL::String::appendf(mainBody, "%s = %s", resultName.c_str(), entry->fStaticFunctionName);
+    *mainBody += append_default_snippet_arguments(entry, curEntryIndex, currentPreLocalName,
+                                                  childOutputVarNames);
+    *mainBody += ";\n";
+#endif  // defined(SK_GRAPHITE_ENABLED) && defined(SK_ENABLE_SKSL)
 }
 
 //--------------------------------------------------------------------------------------------------
@@ -480,33 +548,30 @@ static constexpr char kImageShaderName[] = "sk_compute_coords";
 // Ideally the "compute_coords" code snippet could just take texture and
 // sampler references and do everything. That is going to take more time to figure out though so,
 // for the sake of expediency, we're generating custom code to do the sampling.
-void GenerateImageShaderGlueCode(const std::string& resultName,
-                                 int entryIndex,
+void GenerateImageShaderGlueCode(const SkShaderInfo&,
+                                 const std::string& resultName,
+                                 int* entryIndex,
                                  const SkPaintParamsKey::BlockReader& reader,
-                                 const SkRuntimeEffectDictionary*,
                                  const std::string& priorStageOutputName,
-                                 const std::vector<std::string>& childNames,
+                                 const std::string& currentPreLocalName,
                                  std::string* preamble,
                                  std::string* mainBody,
                                  int indent) {
-    SkASSERT(childNames.empty());
+#if defined(SK_GRAPHITE_ENABLED) && defined(SK_ENABLE_SKSL)
+    std::string samplerVarName = std::string("sampler_") + std::to_string(*entryIndex) + "_0";
 
-    std::string samplerVarName = std::string("sampler_") + std::to_string(entryIndex) + "_0";
-    std::string preLocalMatrixVarName = get_mangled_name("preLocal", entryIndex);
-
-    // Uniform slot 0 is being used for the localMatrix but is handled in
-    // generate_default_before_children_glue_code.
-    std::string subsetName = reader.entry()->getMangledUniformName(1, entryIndex);
-    std::string tmXName = reader.entry()->getMangledUniformName(2, entryIndex);
-    std::string tmYName = reader.entry()->getMangledUniformName(3, entryIndex);
-    std::string imgWidthName = reader.entry()->getMangledUniformName(4, entryIndex);
-    std::string imgHeightName = reader.entry()->getMangledUniformName(5, entryIndex);
+    // Uniform slot 0 is used to make the preLocalMatrix; it's handled in emit_glue_code_for_entry.
+    std::string subsetName = reader.entry()->getMangledUniformName(1, *entryIndex);
+    std::string tmXName = reader.entry()->getMangledUniformName(2, *entryIndex);
+    std::string tmYName = reader.entry()->getMangledUniformName(3, *entryIndex);
+    std::string imgWidthName = reader.entry()->getMangledUniformName(4, *entryIndex);
+    std::string imgHeightName = reader.entry()->getMangledUniformName(5, *entryIndex);
 
     add_indent(mainBody, indent);
     SkSL::String::appendf(mainBody,
                           "float2 coords = %s(%s * dev2LocalUni, %s, %s, %s, %s, %s);",
                           reader.entry()->fStaticFunctionName,
-                          preLocalMatrixVarName.c_str(),
+                          currentPreLocalName.c_str(),
                           subsetName.c_str(),
                           tmXName.c_str(),
                           tmYName.c_str(),
@@ -518,6 +583,7 @@ void GenerateImageShaderGlueCode(const std::string& resultName,
                           "%s = sample(%s, coords);\n",
                           resultName.c_str(),
                           samplerVarName.c_str());
+#endif  // defined(SK_GRAPHITE_ENABLED) && defined(SK_ENABLE_SKSL)
 }
 
 //--------------------------------------------------------------------------------------------------
@@ -600,12 +666,12 @@ private:
 
 #endif
 
-void GenerateRuntimeShaderGlueCode(const std::string& resultName,
-                                   int entryIndex,
+void GenerateRuntimeShaderGlueCode(const SkShaderInfo& shaderInfo,
+                                   const std::string& resultName,
+                                   int* entryIndex,
                                    const SkPaintParamsKey::BlockReader& reader,
-                                   const SkRuntimeEffectDictionary* rteDict,
                                    const std::string& priorStageOutputName,
-                                   const std::vector<std::string>& childOutputVarNames,
+                                   const std::string& currentPreLocalName,
                                    std::string* preamble,
                                    std::string* mainBody,
                                    int indent) {
@@ -613,11 +679,12 @@ void GenerateRuntimeShaderGlueCode(const std::string& resultName,
     const SkShaderSnippet* entry = reader.entry();
 
     // Find this runtime effect in the runtime-effect dictionary.
-    const SkRuntimeEffect* effect = rteDict->find(reader.codeSnippetId());
+    const int codeSnippetId = reader.codeSnippetId();
+    const SkRuntimeEffect* effect = shaderInfo.runtimeEffectDictionary()->find(codeSnippetId);
     SkASSERT(effect);
     const SkSL::Program& program = SkRuntimeEffectPriv::Program(*effect);
 
-    GraphitePipelineCallbacks callbacks{preamble, entryIndex};
+    GraphitePipelineCallbacks callbacks{preamble, *entryIndex};
     SkASSERT(std::string_view(entry->fName) == kRuntimeShaderName);  // the callbacks assume this
     SkSL::PipelineStage::ConvertProgram(program, "coords", "inColor", "half4(1)", &callbacks);
 
@@ -627,19 +694,14 @@ void GenerateRuntimeShaderGlueCode(const std::string& resultName,
     SkASSERT(entry->needsLocalCoords());
     SkASSERT(reader.entry()->fUniforms[0].type() == SkSLType::kFloat4x4);
 
-    std::string preLocalMatrixVarName = get_mangled_name("preLocal", entryIndex);
-
     add_indent(mainBody, indent);
     SkSL::String::appendf(mainBody,
                           "%s = %s_%d((%s * dev2LocalUni * sk_FragCoord).xy, (%s));\n",
                           resultName.c_str(),
-                          entry->fName, entryIndex,
-                          preLocalMatrixVarName.c_str(),
+                          entry->fName, *entryIndex,
+                          currentPreLocalName.c_str(),
                           priorStageOutputName.c_str());
-#else
-    add_indent(mainBody, indent);
-    SkSL::String::appendf(mainBody, "%s = %s;\n", resultName.c_str(), priorStageOutputName.c_str());
-#endif
+#endif  // defined(SK_GRAPHITE_ENABLED) && defined(SK_ENABLE_SKSL)
 }
 
 //--------------------------------------------------------------------------------------------------
@@ -648,16 +710,16 @@ static constexpr char kErrorName[] = "sk_error";
 //--------------------------------------------------------------------------------------------------
 // This method generates the glue code for the case where the SkBlendMode-based blending is
 // handled with fixed function blending.
-void GenerateFixedFunctionBlenderGlueCode(const std::string& resultName,
-                                          int entryIndex,
+void GenerateFixedFunctionBlenderGlueCode(const SkShaderInfo&,
+                                          const std::string& resultName,
+                                          int* entryIndex,
                                           const SkPaintParamsKey::BlockReader& reader,
-                                          const SkRuntimeEffectDictionary*,
                                           const std::string& priorStageOutputName,
-                                          const std::vector<std::string>& childNames,
+                                          const std::string& currentPreLocalName,
                                           std::string* preamble,
                                           std::string* mainBody,
                                           int indent) {
-    SkASSERT(childNames.empty());
+#if defined(SK_GRAPHITE_ENABLED) && defined(SK_ENABLE_SKSL)
     SkASSERT(reader.entry()->fUniforms.empty());
     SkASSERT(reader.numDataPayloadFields() == 0);
 
@@ -668,6 +730,7 @@ void GenerateFixedFunctionBlenderGlueCode(const std::string& resultName,
     *mainBody += "// Fixed-function blending\n";
     add_indent(mainBody, indent);
     SkSL::String::appendf(mainBody, "%s = %s;", resultName.c_str(), priorStageOutputName.c_str());
+#endif  // defined(SK_GRAPHITE_ENABLED) && defined(SK_ENABLE_SKSL)
 }
 
 //--------------------------------------------------------------------------------------------------
@@ -681,20 +744,20 @@ static constexpr char kBlendHelperName[] = "sk_blend";
 // in the shader (i.e., fixed function blending isn't possible).
 // It exists as custom glue code so that we can deal with the dest reads. If that can be
 // standardized (e.g., via a snippets requirement flag) this could be removed.
-void GenerateShaderBasedBlenderGlueCode(const std::string& resultName,
-                                        int entryIndex,
+void GenerateShaderBasedBlenderGlueCode(const SkShaderInfo&,
+                                        const std::string& resultName,
+                                        int* entryIndex,
                                         const SkPaintParamsKey::BlockReader& reader,
-                                        const SkRuntimeEffectDictionary*,
                                         const std::string& priorStageOutputName,
-                                        const std::vector<std::string>& childNames,
+                                        const std::string& currentPreLocalName,
                                         std::string* preamble,
                                         std::string* mainBody,
                                         int indent) {
-    SkASSERT(childNames.empty());
+#if defined(SK_GRAPHITE_ENABLED) && defined(SK_ENABLE_SKSL)
     SkASSERT(reader.entry()->fUniforms.size() == 1);
     SkASSERT(reader.numDataPayloadFields() == 0);
 
-    std::string uniformName = reader.entry()->getMangledUniformName(0, entryIndex);
+    std::string uniformName = reader.entry()->getMangledUniformName(0, *entryIndex);
 
     add_indent(mainBody, indent);
     *mainBody += "// Shader-based blending\n";
@@ -709,6 +772,7 @@ void GenerateShaderBasedBlenderGlueCode(const std::string& resultName,
                           reader.entry()->fStaticFunctionName,
                           uniformName.c_str(),
                           priorStageOutputName.c_str());
+#endif  // defined(SK_GRAPHITE_ENABLED) && defined(SK_ENABLE_SKSL)
 }
 
 //--------------------------------------------------------------------------------------------------
@@ -878,12 +942,6 @@ int SkShaderCodeDictionary::findOrCreateRuntimeEffectSnippet(const SkRuntimeEffe
         return *existingCodeSnippetID;
     }
 
-    // TODO(skia:13405): consider removing these data fields, they don't seem to add value anymore
-    static constexpr DataPayloadField kRuntimeShaderDataPayload[] = {
-            {"runtime effect hash", DataPayloadType::kInt, 1},
-            {"uniform data size (bytes)", DataPayloadType::kInt, 1},
-    };
-
     int newCodeSnippetID = this->addUserDefinedSnippet("RuntimeEffect",
                                                        this->convertUniforms(effect),
                                                        SnippetRequirementFlags::kLocalCoords,
@@ -891,7 +949,7 @@ int SkShaderCodeDictionary::findOrCreateRuntimeEffectSnippet(const SkRuntimeEffe
                                                        kRuntimeShaderName,
                                                        GenerateRuntimeShaderGlueCode,
                                                        /*numChildren=*/0,
-                                                       SkSpan(kRuntimeShaderDataPayload));
+                                                       /*dataPayloadExpectations=*/{});
     fRuntimeEffectMap.set(key, newCodeSnippetID);
     return newCodeSnippetID;
 }
@@ -908,7 +966,7 @@ SkShaderCodeDictionary::SkShaderCodeDictionary() {
             kErrorName,
             GenerateDefaultGlueCode,
             kNoChildren,
-            { }
+            { }      // no data payload
     };
     fBuiltInCodeSnippets[(int) SkBuiltInCodeSnippetID::kSolidColorShader] = {
             "SolidColor",
@@ -918,7 +976,7 @@ SkShaderCodeDictionary::SkShaderCodeDictionary() {
             kSolidShaderName,
             GenerateDefaultGlueCode,
             kNoChildren,
-            { }
+            { }      // no data payload
     };
     fBuiltInCodeSnippets[(int) SkBuiltInCodeSnippetID::kLinearGradientShader4] = {
             "LinearGradient4",
@@ -928,7 +986,7 @@ SkShaderCodeDictionary::SkShaderCodeDictionary() {
             kLinearGradient4Name,
             GenerateDefaultGlueCode,
             kNoChildren,
-            { }
+            { }      // no data payload
     };
     fBuiltInCodeSnippets[(int) SkBuiltInCodeSnippetID::kLinearGradientShader8] = {
             "LinearGradient8",
@@ -938,7 +996,7 @@ SkShaderCodeDictionary::SkShaderCodeDictionary() {
             kLinearGradient8Name,
             GenerateDefaultGlueCode,
             kNoChildren,
-            { }
+            { }      // no data payload
     };
     fBuiltInCodeSnippets[(int) SkBuiltInCodeSnippetID::kRadialGradientShader4] = {
             "RadialGradient4",
@@ -948,7 +1006,7 @@ SkShaderCodeDictionary::SkShaderCodeDictionary() {
             kRadialGradient4Name,
             GenerateDefaultGlueCode,
             kNoChildren,
-            { }
+            { }      // no data payload
     };
     fBuiltInCodeSnippets[(int) SkBuiltInCodeSnippetID::kRadialGradientShader8] = {
             "RadialGradient8",
@@ -958,7 +1016,7 @@ SkShaderCodeDictionary::SkShaderCodeDictionary() {
             kRadialGradient8Name,
             GenerateDefaultGlueCode,
             kNoChildren,
-            { }
+            { }      // no data payload
     };
     fBuiltInCodeSnippets[(int) SkBuiltInCodeSnippetID::kSweepGradientShader4] = {
             "SweepGradient4",
@@ -968,7 +1026,7 @@ SkShaderCodeDictionary::SkShaderCodeDictionary() {
             kSweepGradient4Name,
             GenerateDefaultGlueCode,
             kNoChildren,
-            { }
+            { }      // no data payload
     };
     fBuiltInCodeSnippets[(int) SkBuiltInCodeSnippetID::kSweepGradientShader8] = {
             "SweepGradient8",
@@ -978,7 +1036,7 @@ SkShaderCodeDictionary::SkShaderCodeDictionary() {
             kSweepGradient8Name,
             GenerateDefaultGlueCode,
             kNoChildren,
-            { }
+            { }      // no data payload
     };
     fBuiltInCodeSnippets[(int) SkBuiltInCodeSnippetID::kConicalGradientShader4] = {
             "ConicalGradient4",
@@ -988,7 +1046,7 @@ SkShaderCodeDictionary::SkShaderCodeDictionary() {
             kConicalGradient4Name,
             GenerateDefaultGlueCode,
             kNoChildren,
-            { }
+            { }      // no data payload
     };
     fBuiltInCodeSnippets[(int) SkBuiltInCodeSnippetID::kConicalGradientShader8] = {
             "ConicalGradient8",
@@ -998,7 +1056,7 @@ SkShaderCodeDictionary::SkShaderCodeDictionary() {
             kConicalGradient8Name,
             GenerateDefaultGlueCode,
             kNoChildren,
-            { }
+            { }      // no data payload
     };
     fBuiltInCodeSnippets[(int) SkBuiltInCodeSnippetID::kLocalMatrixShader] = {
             "LocalMatrixShader",
@@ -1006,9 +1064,9 @@ SkShaderCodeDictionary::SkShaderCodeDictionary() {
             SnippetRequirementFlags::kLocalCoords,
             { },     // no samplers
             kLocalMatrixShaderName,
-            GenerateDefaultGlueCode,
+            GenerateDefaultGlueCodeWithChildren,
             kNumLocalMatrixShaderChildren,
-            { }
+            { }      // no data payload
     };
     fBuiltInCodeSnippets[(int) SkBuiltInCodeSnippetID::kImageShader] = {
             "ImageShader",
@@ -1018,7 +1076,7 @@ SkShaderCodeDictionary::SkShaderCodeDictionary() {
             kImageShaderName,
             GenerateImageShaderGlueCode,
             kNoChildren,
-            { }
+            { }      // no data payload
     };
     fBuiltInCodeSnippets[(int) SkBuiltInCodeSnippetID::kBlendShader] = {
             "BlendShader",
@@ -1026,9 +1084,9 @@ SkShaderCodeDictionary::SkShaderCodeDictionary() {
             SnippetRequirementFlags::kNone,
             { },     // no samplers
             kBlendShaderName,
-            GenerateDefaultGlueCode,
+            GenerateDefaultGlueCodeWithChildren,
             kNumBlendShaderChildren,
-            { }
+            { }      // no data payload
     };
     fBuiltInCodeSnippets[(int) SkBuiltInCodeSnippetID::kFixedFunctionBlender] = {
             "FixedFunctionBlender",
@@ -1038,7 +1096,7 @@ SkShaderCodeDictionary::SkShaderCodeDictionary() {
             "FF-blending",  // fixed function blending doesn't use static SkSL
             GenerateFixedFunctionBlenderGlueCode,
             kNoChildren,
-            { }
+            { }      // no data payload
     };
     fBuiltInCodeSnippets[(int) SkBuiltInCodeSnippetID::kShaderBasedBlender] = {
             "ShaderBasedBlender",
@@ -1048,6 +1106,6 @@ SkShaderCodeDictionary::SkShaderCodeDictionary() {
             kBlendHelperName,
             GenerateShaderBasedBlenderGlueCode,
             kNoChildren,
-            { }
+            { }      // no data payload
     };
 }